pip-requirement.txt
Django==1.5.2
django-mptt==0.6.0